services/prediction-service/requirements.txt
flask
flask-cors
pandas
prophet
psycopg2-binary
faker